Fixed on The Null Exceptions/Warnings. Removed The Unused Variables. Soon Will Be Used Tho.
