+++ bin/Debug/net8.0/rcrpt-game-1.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v8.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v8.0": {
+      "rcrpt-game-1/1.0.0": {
+        "runtime": {
+          "rcrpt-game-1.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "rcrpt-game-1/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
